move nydusd client code to daemon package
This is prepared for removing nydussdk package and
make daemon package as a universal exported packge
for other projects.
