+++ pkg/daemon/client_test.go
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package nydussdk
+package daemon
 
 import (
 	"encoding/json"
